+++ clojure-mode.el
@@ -194,7 +194,7 @@ Out-of-the box `clojure-mode' understands lein, boot, gradle,
           (and (listp value)
                (cl-every 'stringp value))))
 
-(defcustom clojure-project-root-function #'clojure-project-root-path
+(defcustom clojure-project-root-function #'clojure-current-project
   "Function to locate clojure project root directory."
   :type 'function
   :risky t
@@ -1664,6 +1664,16 @@ are cached in a buffer local variable (`clojure-cached-project-dir')."
                (not clojure-cached-project-dir))
       (setq clojure-cached-project-dir project-dir))
     project-dir))
+
+(defun clojure-current-project (&optional dir-name)
+  "Return the current project as a cons cell usable by project.el.
+
+Call is delegated down to `clojure-clojure-dir' with
+optional DIR-NAME as argument."
+  (let ((project-dir (clojure-project-dir dir-name)))
+    (if project-dir
+        (cons 'clojure project-dir)
+      nil)))
 
 (defun clojure-project-root-path (&optional dir-name)
   "Return the absolute path to the project's root directory.